deleteOrder now removes every orderMeal with given orderID as well

+++ back-end/src/main/java/pl/pjatk/SOZ_Gastro/Repositories/OrderRepository.java
@@ -3,6 +3,7 @@ package pl.pjatk.SOZ_Gastro.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.pjatk.SOZ_Gastro.ObjectClasses.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Integer>
+public interface OrderRepository extends JpaRepository<Order, Long>
 {
+
 }
